refactor(kraft): Rename `rm` to `remove` to be more consistent
